+++ packages/jsxtreme-markdown/lib/to-component-module.js
@@ -7,6 +7,7 @@ const presetReact = require('babel-preset-react');
 const frontMatter = require('front-matter');
 const pascalCase = require('pascal-case');
 const toJsx = require('./to-jsx');
+const remarkExtractHeadings = require('./remark-extract-headings');
 const defaultTemplate = require('./templates/default');
 
 module.exports = (input, options) => {
@@ -14,17 +15,31 @@ module.exports = (input, options) => {
     {
       name: 'MarkdownReact',
       prependJs: [],
-      precompile: false
+      precompile: false,
+      headings: false,
+      remarkPlugins: []
     },
     options
   );
 
+  if (options.headings) {
+    options.remarkPlugins = [remarkExtractHeadings.plugin].concat(
+      options.remarkPlugins
+    );
+  }
+
   const frontMatterResult = frontMatter(input);
   const jsx = toJsx(frontMatterResult.body, options);
+
+  const extendedFrontMatter = frontMatterResult.attributes;
+  if (options.headings && !extendedFrontMatter.headings) {
+    extendedFrontMatter.headings = remarkExtractHeadings.getHeadings();
+  }
+
   const templateData = {
     name: pascalCase(options.name),
     rawFrontMatter: frontMatterResult.frontmatter,
-    frontMatter: frontMatterResult.attributes,
+    frontMatter: extendedFrontMatter,
     wrapper: frontMatterResult.attributes.wrapper || options.wrapper,
     prependJs: _.union(
       options.prependJs,